+++ base/src/main/java/proguard/analysis/cpa/jvm/domain/reference/JvmReferenceReduceOperator.java
@@ -18,14 +18,12 @@
 
 package proguard.analysis.cpa.jvm.domain.reference;
 
-import java.util.Set;
 import proguard.analysis.cpa.defaults.MapAbstractState;
 import proguard.analysis.cpa.defaults.SetAbstractState;
 import proguard.analysis.cpa.jvm.cfa.nodes.JvmCfaNode;
 import proguard.analysis.cpa.jvm.operators.JvmDefaultReduceOperator;
 import proguard.analysis.cpa.jvm.state.JvmFrameAbstractState;
 import proguard.analysis.cpa.jvm.state.heap.JvmHeapAbstractState;
-import proguard.analysis.cpa.jvm.state.heap.tree.JvmTreeHeapPrincipalAbstractState;
 
 /**
  * This reduce operator behaves the same as the {@link JvmDefaultReduceOperator} but operates on
@@ -48,24 +46,6 @@ public class JvmReferenceReduceOperator
    */
   public JvmReferenceReduceOperator(boolean reduceHeap) {
     super(reduceHeap);
-  }
-
-  // implementations for JvmDefaultReduceOperator
-
-  /**
-   * Performs reduction of the {@link JvmTreeHeapPrincipalAbstractState} keeping just the portion of
-   * the tree rooted at references in static fields and parameters.
-   */
-  @Override
-  protected void reduceHeap(
-      JvmHeapAbstractState<SetAbstractState<Reference>> heap,
-      JvmFrameAbstractState<SetAbstractState<Reference>> reducedFrame,
-      MapAbstractState<String, SetAbstractState<Reference>> reducedStaticFields) {
-    Set<Object> roots = ((JvmTreeHeapPrincipalAbstractState) heap).getStaticCreationReferences();
-    reducedFrame.getLocalVariables().forEach(roots::addAll);
-    reducedStaticFields.values().forEach(roots::addAll);
-
-    heap.reduce(roots);
   }
 
   // implementations for JvmAbstractStateFactory